workers/requirements.txt
# Workers Python - Promo Platform
# ================================

# HTTP requests
requests>=2.31.0
httpx>=0.26.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# OpenAI for copy generation
openai>=1.10.0

# Scheduling
schedule>=1.2.1
APScheduler>=3.10.4

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0

# Async support
asyncio>=3.4.3
aiohttp>=3.9.1

# Logging
loguru>=0.7.2

# Utils
python-dateutil>=2.8.2

# ================================
# Social Media Dispatchers
# ================================

# Twitter/X API
tweepy>=4.14.0

# Telegram Bot API
python-telegram-bot>=20.7

# WhatsApp (via Twilio)
# twilio>=8.11.0

# Facebook (via Meta Graph API)
# facebook-sdk>=3.1.0
